2011-11-13 Roman - !:SYS: Mostly warnings clear caused by option -Wall.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@1643 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/daq/ModBus/modbus_daq.cpp
@@ -1074,9 +1074,9 @@ void TMdPrm::enable( )
 	}catch(TError err) { disable(); throw; }
 
     //> Check for delete DAQ parameter's attributes
-    for(int i_p = 0; i_p < p_el.fldSize(); i_p++)
+    for(int i_p = 0; i_p < (int)p_el.fldSize(); i_p++)
     {
-        int i_l;
+        unsigned i_l;
         for(i_l = 0; i_l < als.size(); i_l++)
             if(p_el.fldAt(i_p).name() == als[i_l])
                 break;
